+++ Marlin/src/feature/digipot/digipot_mcp4451.cpp
@@ -20,13 +20,13 @@
  *
  */
 
-#include "../inc/MarlinConfig.h"
+#include "../../inc/MarlinConfig.h"
 
 #if ENABLED(DIGIPOT_I2C) && DISABLED(DIGIPOT_MCP4018)
 
 #include "Stream.h"
 #include "utility/twi.h"
-#include "Wire.h"
+#include <Wire.h>
 
 // Settings for the I2C based DIGIPOT (MCP4451) on Azteeg X3 Pro
 #if MB(5DPRINT)
@@ -49,15 +49,10 @@ static void i2c_send(const byte addr, const byte a, const byte b) {
 }
 
 // This is for the MCP4451 I2C based digipot
-void digipot_i2c_set_current(uint8_t channel, float current) {
-  current = min((float) max(current, 0.0f), DIGIPOT_I2C_MAX_CURRENT);
+void digipot_i2c_set_current(const uint8_t channel, const float current) {
   // these addresses are specific to Azteeg X3 Pro, can be set to others,
   // In this case first digipot is at address A0=0, A1= 0, second one is at A0=0, A1= 1
-  byte addr = 0x2C; // channel 0-3
-  if (channel >= 4) {
-    addr = 0x2E; // channel 4-7
-    channel -= 4;
-  }
+  const byte addr = channel < 4 ? 0x2C : 0x2E; // channel 0-3 vs 4-7
 
   // Initial setup
   i2c_send(addr, 0x40, 0xFF);
@@ -65,7 +60,7 @@ void digipot_i2c_set_current(uint8_t channel, float current) {
 
   // Set actual wiper value
   byte addresses[4] = { 0x00, 0x10, 0x60, 0x70 };
-  i2c_send(addr, addresses[channel], current_to_wiper(current));
+  i2c_send(addr, addresses[channel & 0x3], current_to_wiper(min((float) max(current, 0.0f), DIGIPOT_I2C_MAX_CURRENT)));
 }
 
 void digipot_i2c_init() {
